Split tasks into appropriate files
